refactor: clean database access methods
This moves all of the database access methods from the GraphQL resolvers to the database package, so they're completely separate and we don't have to pass the database object around so much. This is the first step in reworking the structure of the app to better separate the various parts of the code.