refactor: cleanup providers and resolvers
This completely reworks the way providers and resolvers are handled, and how scanning happens. Resolvers are now handled by a registry, which can pick the proper provider for the libray (and eventually media type). The resolver then gets the basic information of the file (needed for future identification) and creates a blank ItemMetadata with only that info. A provider job is then scheduled, which will handle matching, analyzing and fetching information.