feat(media): add image album support
The logic to create albums is relatively naive. We just check if a directory contains any image. If it does, we consider it to be an album and save it to the DB. Whenever we resolve an image, we figure out which folder it is in and find the relevant album. A thumb for the album is chosen somewhat randomly: the first image of that album to be processed is set as the thumb for the album.