refactor: filesystem scanner
This completely reworks the filesystem scanner to remove the need for a registry and improve performance. Scanners are now independent, but can call each other to modify the list of files, directories and medias.
Golang's walk is replaced by a custom walking algorithm: we simply get the content of the root folder, run them through the scanner, then recursively scan their subfolders.

Currently, only a simple movie scanner is implemented, which just adds all found videos as individual movies (so it doesn't support multi-part movies, etc)